+++ Model/TransportBuilder.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 namespace Yireo\EmailTester2\Model;
 
 use Magento\Framework\Exception\LocalizedException;
+use Magento\Framework\Mail\EmailMessageInterface;
 use Magento\Framework\Mail\MailMessageInterface;
 use Magento\Framework\Mail\Template\TransportBuilder as OriginalTransportBuilder;
 
@@ -24,13 +25,12 @@ use Magento\Framework\Mail\Template\TransportBuilder as OriginalTransportBuilder
 class TransportBuilder extends OriginalTransportBuilder
 {
     /**
-     * @return MailMessageInterface
+     * @return MailMessageInterface|EmailMessageInterface
      * @throws LocalizedException
      */
-    public function getMessage(): MailMessageInterface
+    public function getMessage()
     {
         $this->prepareMessage();
-
         return $this->message;
     }
 }